2023-20b exam modified

+++ 2023-20b/application/app/src/main/java/com/andreeailie/reexam/feature/presentation/components/NetworkObserver.kt
@@ -20,13 +20,13 @@ fun NetworkObserver(snackbarHostState: SnackbarHostState, networkStatusTracker: 
             NetworkStatusTracker.NetworkStatus.Unavailable -> {
                 snackbarHostState.showSnackbar(
                     message = "Network connection lost",
-                    duration = SnackbarDuration.Long
+                    duration = SnackbarDuration.Short
                 )
             }
             NetworkStatusTracker.NetworkStatus.Available -> {
                 snackbarHostState.showSnackbar(
                     message = "Reconnected to the network",
-                    duration = SnackbarDuration.Long
+                    duration = SnackbarDuration.Short
                 )
             }
         }
